Update BootstrapColorPickerPickerField.php
I needed to change the class to extend the FieldTypeBase otherwise the field wouldn't save into the database.
I guess this occurs since Bolt 3.3 because the save mechanism for custom fields had changed. 

Also, the interface FieldTypeBase shouldn't be necessary in this context anymore because the FieldTypeBase itself uses it already.